HrApp/requirements.txt
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
requests==2.28.2
sqlparse==0.4.3
whitenoise==6.3.0
mssql-django==1.2
msal==1.22.0
pandas==2.0.3
requests_oauthlib==1.3.1